Move HTMLWidget to html_widget.h, render HTML in Jupyter when .view() is called
